init event analytics(revenue, attendee, and avg review) by period (daily, weekly, monthly, yearly)

+++ src/main/java/com/riri/eventhop/feature1/reviews/Review.java
@@ -24,7 +24,7 @@ public class Review {
     private User user;
 
     @Column(nullable = false)
-    private Integer rating;
+    private Double rating;
 
     @Column(length = 1000)
     private String comment;
