+++ src/main/java/com/riri/eventhop/feature1/reviews/ReviewResponse.java
@@ -6,7 +6,7 @@ import java.time.Instant;
 @Data
 public class ReviewResponse {
     private String userName;
-    private Integer rating;
+    private Double rating;
     private String comment;
     private String eventName;
     private Instant createdAt;
